Disable auto updating in favor of auto checking
Auto-updating is great. It makes sure users of this config are always on
the latest and greatest and it gives us a hook to pro-actively fix
issues everyone might hit (installing dependencies for go, for example).

But a lot of the times these commits are made in the middle of the day
and people might be working on something and just happened to restart
their editor. Or even in a meeting showing something off to someone
else. At those times having Vim pause and update everything can be a
super frustrating experience.

Moreover, with how slow installing all golang dependencies is these
days, having this happen automatically is quite the surprise if you're
not actually expeciting the update.

This commit embraces the commands `:ConfigUpdate` and `:ConfigUpdate`
and disables the automatic, forced update on boot.

Users will now get a notification to update their configs:

    There is a new version of the nvim config available. Run :ConfigUpdate to update to the latest!

And by default will be directed to use the less intrusive
`:ConfigUpdate` (no `!`) which is much much faster. I expect that lots
of people will end up having to run the `!` version every now and again,
but that can be delayed until they actually have an issue that requires
all dependencies to update.

This direction comes from a timer that runs once on boot and once again
every hour Vim is running and is shown in two ways:

* A warning with the notification above;
* A persistent smaller warning on the bottom right corner of the editor.

[Fixes #20]